// src/shared/types/category.ts
export type Category =
  | 'None'
  | 'General'
  | '.NET'
  | 'CSS'
  | 'Data'
  | 'Design'
  | 'DevOps'
  | 'Go'
  | 'Java'
  | 'Javascript'
  | 'Mobile'
  | 'PHP'
  | 'Python'
  | 'Ruby'
  | 'Security';

export const CategoryArray: Category[] = [
  'None',
  'General',
  '.NET',
  'CSS',
  'Data',
  'Design',
  'DevOps',
  'Go',
  'Java',
  'Javascript',
  'Mobile',
  'PHP',
  'Python',
  'Ruby',
  'Security',
];